refactor: 원 모양의 image view 만드는 코드를 UIImageView extension으로 이동
