+++ App/Sources/Presentation/MyPage/MyPageViewController+.swift
@@ -106,19 +106,4 @@ extension MyPageViewController {
         
         return button
     }
-    
-    func circularImageView(radius: CGFloat) -> UIImageView {
-        let imageView = UIImageView()
-        
-        imageView.layer.cornerRadius = radius
-        imageView.clipsToBounds = true
-        imageView.backgroundColor = .black
-        
-        imageView.snp.makeConstraints {
-            $0.width.equalTo(2 * radius)
-            $0.height.equalTo(2 * radius)
-        }
-        
-        return imageView
-    }
 }
